Rename results page banner logic file
