+++ frontend/src/features/event/results/banner.tsx
